Update HOC consumers js_only
