+++ assets/src/application/ui/input/index.ts
@@ -4,7 +4,6 @@ export { default as DatePicker } from './DateTimePicker/DatePicker';
 export { default as MultiCheck } from './MultiCheck';
 export { default as Number } from './Number';
 export { default as Radio } from './Radio';
-export { default as SearchInput } from './SearchInput';
 export { default as Select } from './Select';
 export { default as Switch } from './Switch';
 export { default as Text } from './Text';
@@ -14,3 +13,6 @@ export { default as TimePicker } from './DateTimePicker/TimePicker';
 
 export * from './EspressoButton';
 export * from './InlineEditInput';
+export * from './SearchInput';
+export * from './SelectInput';
+export * from './SwitchInput';
